backend_auth/requirements.txt
fastapi
uvicorn
sqlmodel
passlib[bcrypt]
pydantic
python-jose[cryptography]
python-multipart

